requirements.txt
Django==5.1.3
django-browser-reload==1.17.0
django-phonenumber-field[phonenumbers]==8.0.0
djlint==1.36.1
python-decouple==3.8
ruff==0.7.3
